# openapi.yaml
openapi: 3.1.0
info:
  title: Agent API
  version: 0.2.0
  description: Unified node agent for Xray, WireGuard, and AmneziaWG (NetinjaBot).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
  - basicAuth: []
paths:
...
  /stats/online:
    get:
      summary: Online users
      parameters:
        - in: query
          name: core
          schema: { type: string }
  /stats/snapshot:
    get:
      summary: Cumulative usage snapshot
      parameters:
        - in: query
          name: core
          schema: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    basicAuth:
      type: http
      scheme: basic